ocultando botones de eliminar y añadiendo validaciones

+++ app/Livewire/Crud/Articulos/EliminarArticuloModal.php
@@ -10,12 +10,27 @@ class EliminarArticuloModal extends Component
     public $item;     // Artículo a eliminar
     public $idModal;  // ID del modal para asociarlo con el botón
 
+    // public function deleteItem()
+    // {
+    //     $this->item->delete();  // Eliminar el artículo
+    //     $this->dispatch('close-modal')->self();  // Emitir evento para cerrar el modal
+    //     $this->dispatch('item-deleted')->to(ContentTable::class);  // Emitir evento para notificar la eliminación
+    // }
+
     public function deleteItem()
     {
-        $this->item->delete();  // Eliminar el artículo
-        $this->dispatch('close-modal')->self();  // Emitir evento para cerrar el modal
-        $this->dispatch('item-deleted')->to(ContentTable::class);  // Emitir evento para notificar la eliminación
+        // Verifica si hay registros en la tabla pivote donacion_articulo con este artículo
+        if ($this->item->donaciones()->exists()) {
+            $this->dispatch('error', 'No se puede eliminar este artículo porque está relacionado con una o más donaciones.');
+            // $this->dispatch('cerrar-modal');
+            return;
+        }
+
+        $this->item->delete();
+        $this->dispatch('cerrar-modal');
+        $this->dispatch('item-deleted');
     }
+
 
     public function mount($parameters)
     {
