+++ app/Livewire/Crud/Donantes/EliminarDonantesModal.php
@@ -9,12 +9,27 @@ class EliminarDonantesModal extends Component
     public $item;
     public $idModal;
 
+    // public function deleteItem()
+    // {
+    //     $this->item->delete();  
+    //     $this->dispatch('cerrar-modal'); 
+    //     $this->dispatch('item-deleted'); 
+    // }
+
     public function deleteItem()
     {
-        $this->item->delete();  
-        $this->dispatch('cerrar-modal'); 
-        $this->dispatch('item-deleted'); 
+        // Validar si el donante tiene donaciones
+        if ($this->item->donaciones()->exists()) {
+            $this->dispatch('error', 'No se puede eliminar este donante porque está relacionado con una o más donaciones.');
+            $this->dispatch('cerrar-modal');
+            return;
+        }
+
+        $this->item->delete();
+        $this->dispatch('cerrar-modal');
+        $this->dispatch('item-deleted');
     }
+
 
     public function mount($parameters)
     {
